+++ CHANGELOG.md
@@ -44,6 +44,7 @@
 - Fix precision problem in patterns when overzoomed in OpenGL ES devices.
 - Fix padding-top of the popup to improve readability of popup text (#354).
 - Fix GeoJSONSource#loaded sometimes returning true while there are still pending loads (#669)
+- Fix MapDataEvent#isSourceLoaded being true in GeoJSONSource "dataloading" event handlers (#694)
 
 ## 1.15.2